+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Pearl changelog
 
 ## [Unreleased]
 #### Added
-
+- Corrupted blobs count calculation (#464)
 
 #### Changed
 - Save delete records in index (#207)